refactor(discord): move and modularize discord event handlers - index.js:   - remove inline command registration and event handlers   - import and use setupDiscordHandlers - discord/index.js:   - create setupDiscordHandlers function   - orchestrate loading of discord-related modules - discord/clientEvents.js:   - handle generic discord client events (error, disconnect, reconnect, etc.) - discord/commands.js:   - define and register opt-in and opt-out slash commands - discord/presenceHandlers.js:   - handle presenceUpdate and userUpdate events   - debounce and send activity/user data - discord/slashHandlers.js:   - handle interactionCreate event   - process opt-in and opt-out slash commands
